pyproject.toml
[tool.poetry]
name = "korat"
version = "0.0.1"
description = "Friendly stats bot for hound.fm"
authors = ["btzr-io <btzr.io@gmail.com>"]
license = "MIT"
packages = [
    { include = "main.py", from = "src" },
]

[tool.poetry.scripts]
korat = "main:run"

[tool.poetry.dependencies]
python = "^3.8"
black = "^20.8b1"
eland = "^7.10.1-beta.1"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
